Wow
Heavy rewrite of the app going on lol
> Note: When will it be done is unknown. I just code it
whenever I get some time (or rather I want some change from
school stuff)
The goal with this intense rewrite is to shift to more
kotlin-based ecosystem and libraries and be neatly
programmed.

Although the app was decently commented and run, it wasn't
the best code nor the best performing app. (and I cut some
corners by entirely removing the app from home screen
and making it a quick setting tile only app)

At the end of this rewrite, this should be a responsive and
usable app.

Also, this app uses more modern technologies like Material 3
design.